Refactor reducers to a specific folder
